Add household raw data
