+++ retail-insights/src/pages/CombinedDataPage.js
@@ -1,128 +1,162 @@
-import React, { useState } from 'react';
+// src/pages/CombinedDataPage.js
+import React, { useState, useEffect } from 'react'
+import axios from 'axios'
+import hh10CsvUrl from '../data/hh10.csv'
 
-const CombinedDataPage = () => {
-  const data = [];
+const averageBasketMetrics = [
+  { metric: 'Average Basket Price', value: '$57.30' },
+  { metric: 'Average Basket Size (units)', value: '11.85' }
+]
+
+const tableWrapperStyle = {
+  display: 'flex',
+  justifyContent: 'center',
+  margin: '2em 0',
+}
+
+const tableStyle = {
+  borderCollapse: 'collapse',
+  margin: '0 auto',
+  textAlign: 'center',
+}
+
+const thTdStyle = {
+  border: '1px solid #999',
+  padding: '0.5em 1em',
+  textAlign: 'center',
+}
+
+function CombinedDataPage() {
+  const [data, setData]           = useState([])
   const [household, setHousehold] = useState('')
-  const [record, setRecord]     = useState(null);
+  const [record, setRecord]       = useState(null)
+  const [csvLines, setCsvLines]   = useState([])
 
-  
+  useEffect(() => {
+    axios.get('/data/combined_data.json')
+      .then(res => setData(res.data))
+      .catch(() => setData([]))
+  }, [])
+
+  useEffect(() => {
+    fetch(hh10CsvUrl)
+      .then(res => res.text())
+      .then(text => setCsvLines(text.trim().split('\n')))
+      .catch(err => console.error(err))
+  }, [])
+
   const handleSubmit = e => {
     e.preventDefault()
     const num = parseInt(household, 10)
-    const found = data.filter(item => item.HSHD_NUM === num)
+    const found = data.find(item => item.HSHD_NUM === num)
     setRecord(found || { notFound: true })
   }
 
-  return (
-    <div className="p-6">
-      <h2 className="text-2xl font-bold mb-4">Lookup Household Record</h2>
+  const renderTable = (captionText, headers, rows) => (
+    <div style={tableWrapperStyle}>
+      <table style={tableStyle}>
+        <caption style={{ fontSize: '1.5em', fontWeight: 'bold', marginBottom: '0.5em' }}>
+          {captionText}
+        </caption>
+        <thead>
+          <tr>
+            {headers.map(h => (
+              <th key={h} style={{ ...thTdStyle, background: '#eee' }}>
+                {h}
+              </th>
+            ))}
+          </tr>
+        </thead>
+        <tbody>
+          {rows.map((row, i) => (
+            <tr key={i}>
+              {row.map((cell, j) => (
+                <td key={j} style={thTdStyle}>
+                  {cell}
+                </td>
+              ))}
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  )
 
-      <form onSubmit={handleSubmit} className="mb-6">
-        <label className="block mb-2">
+  // Summary table data
+  const summaryHeaders = ['Metric', 'Value']
+  const summaryRows = averageBasketMetrics.map(({ metric, value }) => [metric, value])
+
+  // Raw CSV table data
+  let csvTable = null
+  if (csvLines.length > 1) {
+    const [headerLine, ...dataLines] = csvLines
+    const csvHeaders = headerLine.split(',')
+    const csvRows = dataLines.map(line => line.split(','))
+    csvTable = renderTable('Household 10 Raw Data', csvHeaders, csvRows)
+  }
+
+  return (
+    <div style={{ padding: '1em' }}>
+      <h2 style={{ fontSize: '1.5em', marginBottom: '0.5em' }}>
+        Lookup Household Record
+      </h2>
+      <form onSubmit={handleSubmit} style={{ marginBottom: '1em' }}>
+        <label style={{ display: 'block', marginBottom: '0.5em' }}>
           Household Number:{' '}
           <input
             type="number"
             value={household}
             onChange={e => setHousehold(e.target.value)}
-            className="border p-2 rounded"
+            style={{ border: '1px solid #ccc', padding: '0.5em', borderRadius: '4px' }}
             required
           />
         </label>
         <button
           type="submit"
-          className="px-4 py-2 bg-blue-600 text-white rounded"
+          style={{
+            padding: '0.5em 1em',
+            background: '#2563eb',
+            color: 'white',
+            border: 'none',
+            borderRadius: '4px',
+            cursor: 'pointer',
+          }}
         >
           Search
         </button>
       </form>
 
       {record && record.notFound && (
-        <p className="text-red-600">
+        <p style={{ color: '#b91c1c' }}>
           No record found for household #{household}
         </p>
       )}
 
       {record && !record.notFound && (
-        <div className="overflow-auto">
-          <table className="min-w-full table-auto border-collapse">
+        <div style={tableWrapperStyle}>
+          <table style={tableStyle}>
             <thead>
               <tr>
-                <th className="border px-4 py-2">Field</th>
-                <th className="border px-4 py-2">Value</th>
+                <th style={{ ...thTdStyle, background: '#eee' }}>Field</th>
+                <th style={{ ...thTdStyle, background: '#eee' }}>Value</th>
               </tr>
             </thead>
             <tbody>
               {Object.entries(record).map(([key, value]) => (
                 <tr key={key}>
-                  <td className="border px-4 py-2 font-medium">{key}</td>
-                  <td className="border px-4 py-2">
-                    {value != null ? value.toString() : ''}
-                  </td>
+                  <td style={thTdStyle}>{key}</td>
+                  <td style={thTdStyle}>{value != null ? value.toString() : ''}</td>
                 </tr>
               ))}
             </tbody>
           </table>
         </div>
       )}
+
+      {renderTable('Household 10 data', summaryHeaders, summaryRows)}
+      {csvTable}
     </div>
   )
 }
-
-
-const averageBasketMetrics = [
-  { metric: 'Average Basket Price', value: '$57.30' },
-  { metric: 'Average Basket Size (units)', value: '11.85' }
-];
-
-
-function renderTable(data) {
-  
-  const table = document.createElement('table');
-  table.style.borderCollapse = 'collapse';
-  table.style.margin = '2em auto';         
-  table.style.textAlign = 'left';           
-
-
-  const caption = document.createElement('caption');
-  caption.textContent = 'Household 10 data';
-  caption.style.fontSize = '1.5em';
-  caption.style.fontWeight = 'bold';
-  caption.style.marginBottom = '0.5em';
-  table.appendChild(caption);
-
-
-  const headerRow = document.createElement('tr');
-  ['Metric', 'Value'].forEach(text => {
-    const th = document.createElement('th');
-    th.textContent = text;
-    th.style.border = '1px solid #999';
-    th.style.padding = '0.5em 1em';
-    th.style.background = '#eee';
-    headerRow.appendChild(th);
-  });
-  table.appendChild(headerRow);
-
-
-  data.forEach(({ metric, value }) => {
-    const row = document.createElement('tr');
-    [metric, value].forEach(text => {
-      const td = document.createElement('td');
-      td.textContent = text;
-      td.style.border = '1px solid #999';
-      td.style.padding = '0.5em 1em';
-      row.appendChild(td);
-    });
-    table.appendChild(row);
-  });
-
-
-  document.body.appendChild(table);
-}
-
-
-document.addEventListener('DOMContentLoaded', () => {
-  renderTable(averageBasketMetrics);
-});
-
 
 export default CombinedDataPage
